Fix Add to Cart Functionality

+++ app/controllers/carts_controller.rb
@@ -8,39 +8,57 @@ class CartsController < ApplicationController
       @cart = current_user.cart || current_user.create_cart
       @cart_items = @cart.cart_items
     else
-      session[:cart] ||= {}  # Ensure session cart exists
+      session[:cart] ||= {}
       @cart_items = session[:cart].map do |product_id, quantity|
         product = Product.find_by(id: product_id)
         OpenStruct.new(product: product, quantity: quantity, id: product_id) if product
       end.compact
     end
   end
-  
 
   def add_item
+    product = Product.find_by(id: params[:product_id])
+
+    if product.nil?
+      redirect_to cart_path, alert: "Product not found!" and return
+    end
+
+    quantity = params[:quantity].to_i
+    if quantity < 1
+      redirect_to cart_path, alert: "Invalid quantity!" and return
+    end
+
     if user_signed_in?
       @cart = current_user.cart || current_user.create_cart
-      @cart_item = @cart.cart_items.find_or_initialize_by(product_id: params[:product_id])
-      @cart_item.quantity += params[:quantity].to_i
+      @cart_item = @cart.cart_items.find_or_initialize_by(product_id: product.id)
+      @cart_item.quantity += quantity
       @cart_item.save
     else
       session[:cart] ||= {}
-      session[:cart][params[:product_id]] ||= 0
-      session[:cart][params[:product_id]] += params[:quantity].to_i
+      session[:cart][product.id.to_s] ||= 0
+      session[:cart][product.id.to_s] += quantity
     end
+
     redirect_to cart_path, notice: "Product added to cart!"
   end
 
   def remove_item
     if user_signed_in?
-      @cart_item = CartItem.find(params[:id])
-      @cart_item.destroy
+      cart_item = CartItem.find_by(id: params[:id])
+      if cart_item
+        cart_item.destroy
+        redirect_to cart_path, notice: "Item removed from cart!"
+      else
+        redirect_to cart_path, alert: "Item not found in cart!"
+      end
     else
-      session[:cart].delete(params[:product_id]) # Use product_id instead of id
+      session[:cart] ||= {}
+      if session[:cart].key?(params[:product_id])
+        session[:cart].delete(params[:product_id])
+        redirect_to cart_path, notice: "Item removed from cart!"
+      else
+        redirect_to cart_path, alert: "Item not found in cart!"
+      end
     end
-  
-    redirect_to cart_path, notice: "Product removed from cart!"
   end
-  
-  
 end
